+++ ipc/schema/packageJson.ts
@@ -10,6 +10,20 @@ import { knownDevicePcas } from '../device';
 import { nrfModules, semver } from '../MetaFiles';
 import { parseWithPrettifiedErrorMessage } from './parseJson';
 
+const packageJson = z.object({
+    name: z.string(),
+    version: semver,
+
+    displayName: z.string().optional(),
+});
+
+export type PackageJson = z.infer<typeof packageJson>;
+
+export const parsePackageJson =
+    parseWithPrettifiedErrorMessage<PackageJson>(packageJson);
+
+// Apps have more required fields in their package.json
+
 const nrfConnectForDesktop = z.object({
     supportedDevices: z.enum(knownDevicePcas).array().nonempty().optional(),
     nrfutil: nrfModules.optional(),
@@ -22,12 +36,7 @@ const engines = recordOfOptionalStrings.and(
     z.object({ nrfconnect: z.string() })
 );
 
-const packageJson = z.object({
-    name: z.string(),
-    version: semver,
-
-    author: z.string().optional(),
-    bin: z.string().or(recordOfOptionalStrings).optional(),
+const packageJsonApp = packageJson.extend({
     dependencies: recordOfOptionalStrings.optional(),
     description: z.string(),
     homepage: z.string().url().optional(),
@@ -36,8 +45,6 @@ const packageJson = z.object({
     engines,
     nrfConnectForDesktop,
     files: z.string().array().optional(),
-    license: z.string().optional(),
-    main: z.string().optional(),
     peerDependencies: recordOfOptionalStrings.optional(),
     repository: z
         .object({
@@ -45,25 +52,22 @@ const packageJson = z.object({
             url: z.string().url(),
         })
         .optional(),
-    scripts: recordOfOptionalStrings.optional(),
 });
 
-export type PackageJson = z.infer<typeof packageJson>;
+export type PackageJsonApp = z.infer<typeof packageJsonApp>;
 
-export const parsePackageJson =
-    parseWithPrettifiedErrorMessage<PackageJson>(packageJson);
+export const parsePackageJsonApp =
+    parseWithPrettifiedErrorMessage<PackageJsonApp>(packageJsonApp);
 
 // In the launcher we want to handle that the whole nrfConnectForDesktop may be missing
 // and the html in it can also be undefined, so there we need to use this legacy variant
-const legacyPackageJson = packageJson.merge(
-    z.object({
-        nrfConnectForDesktop: nrfConnectForDesktop
-            .partial({ html: true })
-            .optional(),
-    })
-);
+const packageJsonLegacyApp = packageJsonApp.extend({
+    nrfConnectForDesktop: nrfConnectForDesktop
+        .partial({ html: true })
+        .optional(),
+});
 
-export type LegacyPackageJson = z.infer<typeof legacyPackageJson>;
+export type PackageJsonLegacyApp = z.infer<typeof packageJsonLegacyApp>;
 
-export const parseLegacyPackageJson =
-    parseWithPrettifiedErrorMessage<LegacyPackageJson>(legacyPackageJson);
+export const parsePackageJsonLegacyApp =
+    parseWithPrettifiedErrorMessage<PackageJsonLegacyApp>(packageJsonLegacyApp);