+++ mocks/packageJsonMock.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-export default () => ({
+export const packageJson = () => ({
     name: 'mocked-test-app',
     version: '27.6.72',
 });