+++ nrfutil/sandbox.ts
@@ -10,6 +10,7 @@ import os from 'os';
 import path from 'path';
 
 import describeError from '../src/logging/describeError';
+import usageData from '../src/utils/usageData';
 import { versionToInstall } from './moduleVersion';
 import { getNrfutilLogger } from './nrfutilLogger';
 import {
@@ -266,7 +267,7 @@ export class NrfutilSandbox {
             }
 
             error.message = error.message.replaceAll('Error: ', '');
-            getNrfutilLogger()?.error(
+            usageData.sendErrorReport(
                 `${
                     pid && this.logLevel === 'trace' ? `[PID:${pid}] ` : ''
                 }${describeError(error)}`
@@ -301,6 +302,7 @@ export class NrfutilSandbox {
     ) =>
         new Promise<void>((resolve, reject) => {
             let aborting = false;
+            usageData.sendUsageData(`running nrfutil ${command}`, { args });
             const nrfutil = spawn(
                 path.join(this.baseDir, 'nrfutil'),
                 [
@@ -407,6 +409,7 @@ export class NrfutilSandbox {
                 closedHandlers.forEach(callback => callback());
             })
             .catch(error => {
+                usageData.sendErrorReport(describeError(error));
                 running = false;
                 closedHandlers.forEach(callback => callback(error));
             });