+++ src/utils/systemReport.ts
@@ -8,7 +8,7 @@ import fs from 'fs';
 import { EOL } from 'os';
 import path from 'path';
 import pretty from 'prettysize';
-import type Systeminformation from 'systeminformation';
+import si from 'systeminformation';
 
 import NrfutilDeviceLib from '../../nrfutil/device/device';
 import {
@@ -25,8 +25,6 @@ import { getAppDataDir } from './appDirs';
 import { openFile } from './open';
 
 const generalInfoReport = async () => {
-    // eslint-disable-next-line global-require
-    const si = require('systeminformation') as typeof Systeminformation;
     const [
         { manufacturer, model },
         { vendor, version },